src/llm/feature_developer.py
from PyQt6.QtCore import QThread, pyqtSignal
import openai
import anthropic
from transformers import pipeline
import ast

class FeatureDeveloper(QThread):
    development_progress = pyqtSignal(str)
    development_complete = pyqtSignal(dict)

    def __init__(self, feature, llm_provider, api_key):
        super().__init__()
        self.feature = feature
        self.llm_provider = llm_provider
        self.api_key = api_key

    def run(self):
        self.development_progress.emit(f"Developing feature: {self.feature['name']}...")
        implementation = self.develop_feature()
        self.development_complete.emit(implementation)

    def develop_feature(self):
        if self.feature['name'] == "Self Healing":
            return self.generate_self_healing_suggestions()
        else:
            prompt = (
                f"Develop a Python implementation for the following feature to be added to a codebase analyzer application:\n\n"
                f"Feature Name: {self.feature['name']}\n"
                f"Description: {self.feature['description']}\n\n"
                f"Provide the implementation as a new method to be added to the MainWindow class. "
                f"Also include any necessary imports or new class definitions. "
                f"Format your response as a JSON object with two keys: 'method' (the new method to be added to MainWindow) "
                f"and 'additions' (any new imports or class definitions to be added at the top of the file)."
            )

            if self.llm_provider == "OpenAI":
                return self.develop_with_openai(prompt)
            elif self.llm_provider == "Claude":
                return self.develop_with_claude(prompt)
            else:
                return self.develop_with_huggingface(prompt)

    def generate_self_healing_suggestions(self):
        prompt = (
            "Analyze the following Python codebase and suggest improvements for each file. "
            "Focus on code quality, performance, and best practices. "
            "Provide specific code changes for each suggestion. "
            "Format your response as a JSON object where keys are file paths and values are the improved code content."
        )

        # Here we would normally pass the entire codebase, but for brevity, let's assume we're working with a single file
        sample_code = '''
def complex_function(x, y):
    z = 0
    for i in range(x):
        for j in range(y):
            z += i * j
    return z

def unused_function():
    pass

global_var = 42

class PoorlyNamedClass:
    def __init__(self):
        self.x = 10
        self.y = 20
        self.z = 30
    
    def poorly_named_method(self):
        return self.x + self.y + self.z
'''
        prompt += f"\n\nCodebase:\n{sample_code}"

        if self.llm_provider == "OpenAI":
            return self.develop_with_openai(prompt)
        elif self.llm_provider == "Claude":
            return self.develop_with_claude(prompt)
        else:
            return self.develop_with_huggingface(prompt)

    def develop_with_openai(self, prompt):
        openai.api_key = self.api_key
        response = openai.ChatCompletion.create(
            model="gpt-3.5-turbo",
            messages=[
                {"role": "system", "content": "You are a skilled Python developer that implements new features for software applications."},
                {"role": "user", "content": prompt}
            ]
        )
        return self.parse_and_validate_code(response.choices[0].message.content.strip())

    def develop_with_claude(self, prompt):
        client = anthropic.Client(api_key=self.api_key)
        response = client.completion(
            prompt=f"Human: {prompt}\n\nAssistant:",
            model="claude-2",
            max_tokens_to_sample=2000,
        )
        return self.parse_and_validate_code(response.completion.strip())

    def develop_with_huggingface(self, prompt):
        generator = pipeline('text-generation', model='gpt2')
        response = generator(prompt, max_length=2000, num_return_sequences=1)[0]['generated_text']
        # Note: This is a simplification. In practice, you'd need more sophisticated parsing for HuggingFace output.
        return {"method": "def placeholder_method(self):\n    pass", "additions": ""}

    def parse_and_validate_code(self, response):
        try:
            code_dict = ast.literal_eval(response)
            # Basic validation
            if not isinstance(code_dict, dict) or 'method' not in code_dict or 'additions' not in code_dict:
                raise ValueError("Invalid response format")
            
            # Attempt to parse the method to ensure it's valid Python code
            ast.parse(code_dict['method'])
            
            return code_dict
        except (SyntaxError, ValueError) as e:
            self.development_progress.emit(f"Error in generated code: {str(e)}")
            return {"method": "def error_method(self):\n    pass", "additions": "# Error in code generation"}